+++ graphein/protein/config.py
@@ -12,7 +12,7 @@ from pathlib import Path
 from typing import Any, Callable, List, Optional, Union
 
 from deepdiff import DeepDiff
-from pydantic import BaseModel
+from pydantic import BaseModel, validator
 from typing_extensions import Literal
 
 from graphein.protein.edges.distance import add_peptide_bonds
@@ -103,6 +103,14 @@ GraphAtoms = Literal[
 GranularityOpts = Literal["atom", "centroids"]
 """Allowable granularity options for nodes in the graph."""
 
+AltLocsOpts = Union[
+    bool,
+    Literal[
+        "max_occupancy", "min_occupancy", "first", "last", "exclude", "include"
+    ],
+]
+"""Allowable altlocs options for alternative locations handling."""
+
 
 class ProteinGraphConfig(BaseModel):
     """
@@ -131,6 +139,12 @@ class ProteinGraphConfig(BaseModel):
     :type keep_hets: List[str]
     :param insertions: Controls whether or not insertions are allowed.
     :type insertions: bool
+    :param alt_locs: Controls whether or not alternative locations are allowed. The supported values are
+        ``"max_occupancy"``, ``"min_occupancy"``, ``"first"``, ``"last"``, ``"exclude"``. First two will leave altlocs
+        with the highest/lowest occupancies, next two will leave first/last in the PDB file ordering. The ``"exclude"``
+        value will drop them entirely and ``"include"`` leave all of them. Additionally, boolean values are the aliases
+        for the latest options. Default is ``"max_occupancy"``.
+    :type alt_locs: AltLocsOpts
     :param pdb_dir: Specifies path to download protein structures into.
     :type pdb_dir: pathlib.Path. Optional.
     :param verbose: Specifies verbosity of graph creation process.
@@ -173,7 +187,8 @@ class ProteinGraphConfig(BaseModel):
 
     granularity: Union[GraphAtoms, GranularityOpts] = "CA"
     keep_hets: List[str] = []
-    insertions: bool = False
+    insertions: bool = True
+    alt_locs: AltLocsOpts = "max_occupancy"
     pdb_dir: Optional[Path] = None
     verbose: bool = False
     exclude_waters: bool = True
@@ -193,6 +208,15 @@ class ProteinGraphConfig(BaseModel):
     # External Dependency configs
     get_contacts_config: Optional[GetContactsConfig] = None
     dssp_config: Optional[DSSPConfig] = None
+
+    @validator("alt_locs")
+    def convert_alt_locs_aliases(cls, v):
+        if v is True:
+            return "include"
+        elif v is False:
+            return "exclude"
+        else:
+            return v
 
     def __eq__(self, other: Any) -> bool:
         """