+++ tests/protein/test_graphs.py
@@ -5,6 +5,7 @@ from pathlib import Path
 
 import networkx as nx
 import pandas as pd
+import numpy as np
 import pytest
 
 from graphein.protein.config import DSSPConfig, ProteinGraphConfig
@@ -360,11 +361,12 @@ def test_graph_sequence_feature():
         assert g_atom.graph[f"sequence_{c}"] == g_res.graph[f"sequence_{c}"]
 
 
-def test_insertion_handling():
+def test_insertion_and_alt_loc_handling():
     configs = {
         "granularity": "CA",
         "keep_hets": [],
         "insertions": False,
+        "alt_locs": "max_occupancy",
         "verbose": False,
         "node_metadata_functions": [meiler_embedding, expasy_protein_scale],
         "edge_construction_functions": [
@@ -386,6 +388,73 @@ def test_insertion_handling():
         g.graph["sequence_C"]
     ) + len(g.graph["sequence_D"]) + len(g.graph["sequence_E"]) == len(g)
     assert g.graph["coords"].shape[0] == len(g)
+
+
+def test_alt_loc_exclusion():
+    configs = {
+        "granularity": "CA",
+        "keep_hets": [],
+        "insertions": True,
+        "alt_locs": "max_occupancy",
+        "verbose": False,
+        "node_metadata_functions": [meiler_embedding, expasy_protein_scale],
+        "edge_construction_functions": [
+            add_peptide_bonds,
+            add_hydrogen_bond_interactions,
+            add_ionic_interactions,
+            add_aromatic_sulphur_interactions,
+            add_hydrophobic_interactions,
+            add_cation_pi_interactions,
+        ],
+    }
+
+    config = ProteinGraphConfig(**configs)
+
+    # This is a PDB with three altlocs
+    g = construct_graph(config=config, pdb_code="2VVI")
+
+    # Test altlocs are dropped
+    assert len(set(g.nodes())) == len(g.nodes())
+
+    # Test the correct one is left
+    for opt, expected_coords, node_id in (
+        ("max_occupancy", [5.850, -9.326, -42.884], "A:CYS:195:A"),
+        ("min_occupancy", [5.864, -9.355, -42.943], "A:CYS:195:B"),
+        ("first", [5.850, -9.326, -42.884], "A:CYS:195:A"),
+        ("last", [5.864, -9.355, -42.943], "A:CYS:195:B"),
+    ):
+        config.alt_locs = opt
+        g = construct_graph(config=config, pdb_code="2VVI")
+        assert np.array_equal(g.nodes[node_id]["coords"], expected_coords)
+
+
+def test_alt_loc_inclusion():
+    configs = {
+        "granularity": "CA",
+        "keep_hets": [],
+        "insertions": False,
+        "alt_locs": True,
+        "verbose": False,
+        "node_metadata_functions": [meiler_embedding, expasy_protein_scale],
+        "edge_construction_functions": [
+            add_peptide_bonds,
+            add_hydrogen_bond_interactions,
+            add_ionic_interactions,
+            add_aromatic_sulphur_interactions,
+            add_hydrophobic_interactions,
+            add_cation_pi_interactions,
+        ],
+    }
+
+    config = ProteinGraphConfig(**configs)
+
+    # This is a PDB with an altloc leading to different residues
+    g = construct_graph(config=config, pdb_code="1ALX")
+
+    # Test both are present
+    assert "A:TYR:11:A" in g.nodes() and "A:TRP:11:B" in g.nodes()
+
+    # TODO Test on other PDBs where altlocs are of the same residues
 
 
 def test_edges_do_not_add_nodes_for_chain_subset():